Agno(Phidata) / Day 01/requirements.txt
agno
ollama
groq
python-dotenv
duckduckgo-search
streamlit